fix broken test from develop

+++ src/libais/ais.h
@@ -853,9 +853,9 @@ ostream& operator<< (ostream &o, const Ais6_1_40 &msg);
 // Number of persons on board.  ITU 1371-1
 class Ais6_235_10 : public Ais6 {
  public:
-  int ana_int;
-  int ana_ext1;
-  int ana_ext2;
+  float ana_int;
+  float ana_ext1;
+  float ana_ext2;
   int racon;
   int light;
   bool health;
